+++ myems-admin/app/controllers/settings/command/command.master.controller.js
@@ -1,0 +1,9 @@
+'use strict';
+
+app.controller('CommandMasterController', function($scope) {
+
+	$scope.$on('handleEmitCommandChanged', function(event) {
+		$scope.$broadcast('handleBroadcastCommandChanged');
+	});
+
+});